refactor(transit): centralize itinerary processing
